SkyLedge/OBD_Logger/requirements.txt
# Server
fastapi
uvicorn[standard]
python-multipart
jinja2

# Data
pandas
numpy
scikit-learn

# ML Models
xgboost
joblib

# Drive
gspread
oauth2client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client

# Database
pymongo
google-cloud-storage
firebase-admin

# Visualize
matplotlib
seaborn

# HuggingFace
huggingface_hub==0.25.2

# Additional dependencies for RLHF training
pyarrow  # For parquet file support
